#48 Re-wrote audiohandler into audio.java and audiohandler.java so as to enable multiple audio instances at once. Also enables programmatic control.
Combined audio() and createAudio() to one method, they were
unnecessarily separate. Also, re-organised code for readability.


<ca718@york.ac.uk>
